convex-hull-after-effects-expression
An After Effects expression that calculates the minimum enclosing shape given a set of 2D points. Meant to be used on a mask path or Shape Layer path.

A note: I think the expression adds the left-most hull point twice-- once at the beginning and once at the end. Could probably be refactored to deal with that redundancy.